fix: padding issues

+++ inc/fmt.h
@@ -6,7 +6,7 @@
 /*   By: nmartins <nmartins@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/05/22 16:33:57 by nmartins       #+#    #+#                */
-/*   Updated: 2019/06/26 18:57:02 by nmartins      ########   odam.nl         */
+/*   Updated: 2019/07/05 16:16:00 by nmartins      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,24 +52,24 @@ void				intern_fmt_puthex(
 						t_writer *writer,
 						unsigned long long value,
 						int caps);
-void				intern_fmt_pad(t_writer *writer, char c, size_t amt);
+void				intern_fmt_pad(t_writer *writer, char c, int amt);
 char				intern_to_hex(char num, int caps);
 void				intern_pop_wildcards(t_token *token, va_list vlist);
 void				intern_fmt_pad_auto(
 						t_writer *writer,
 						char c,
-						size_t pad_amount,
-						size_t length);
+						int pad_amount,
+						int length);
 void				intern_fmt_pad_left(
 						t_writer *writer,
 						t_token *token,
 						char c,
-						size_t length);
+						int length);
 void				intern_fmt_pad_right(
 						t_writer *writer,
 						t_token *token,
 						char c,
-						size_t length);
+						int length);
 int					intern_hex_size(long long n);
 void				intern_fmt_color(t_writer *writer, t_color color);
 size_t				intern_ntoa(char *buf_out, t_number number, int caps);
